[Silver III] Title: 주유소, Time: 304 ms, Memory: 32500 KB, Score: 17 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -7,18 +7,20 @@ public class Main {
         int N = Integer.parseInt(br.readLine()); // 도시의 개수
 
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int[] dist = new int[N-1]; // 도로의 길이
+        long[] dist = new long[N-1]; // 도로의 길이
         for (int i = 0; i < N-1; i++) {
-            dist[i] = Integer.parseInt(st.nextToken());
+            dist[i] = Long.parseLong(st.nextToken());
         }
 
         st = new StringTokenizer(br.readLine());
-        int[] price = new int[N]; // 주유소의 리터당 가격
+        long[] price = new long[N]; // 주유소의 리터당 가격
         for (int i = 0; i < N; i++) {
-            price[i] = Integer.parseInt(st.nextToken());
+            price[i] = Long.parseLong(st.nextToken());
         }
 
-        int[] cost = new int[N];
+        // 도시 별 기름 가격을 내림차순으로 만들기
+        // ex. [8, 9, 5, 4, 2, 7, 1] → [8, 8, 5, 4, 2, 2, 1]
+        long[] cost = new long[N];
         cost[0] = price[0];
         for (int i = 1; i < N; i++) {
             if (price[i-1] < price[i]) cost[i] = cost[i-1];
@@ -27,7 +29,7 @@ public class Main {
 
         long total = 0;
         for (int i = 0; i < N-1; i++) {
-            total += (long) cost[i] * dist[i];
+            total += cost[i] * dist[i];
         }
 
         System.out.println(total);
